move pgconfig
