+++ .vscode/settings.json
@@ -13,5 +13,6 @@
             "LICENSE": true,
             ".git/logs": true,
       },
+      "editor.tabSize": 4,
 }
